Refactor commands into new structure
