+++ src/snowcli/cli/snowpark/__init__.py
@@ -4,9 +4,12 @@ from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
 from snowcli.cli.snowpark.function import app as function_app
 from snowcli.cli.snowpark.package import app as package_app
 from snowcli.cli.snowpark.procedure import app as procedure_app
-from snowcli.cli.snowpark.cp import app as compute_pools_app, app_cp as cp_app
-from snowcli.cli.snowpark.services import app as services_app
-from snowcli.cli.snowpark.jobs import app as jobs_app
+from snowcli.cli.snowpark.compute_pool.commands import (
+    app as compute_pools_app,
+    app_cp as cp_app,
+)
+from snowcli.cli.snowpark.services.commands import app as services_app
+from snowcli.cli.snowpark.jobs.commands import app as jobs_app
 from snowcli.cli.snowpark.registry import app as registry_app
 
 app = typer.Typer(
