+++ src/snowcli/cli/snowpark/compute_pool/manager.py
@@ -1,0 +1,25 @@
+from snowcli.cli.common.sql_execution import SqlExecutionMixin
+
+
+class ComputePoolManager(SqlExecutionMixin):
+    def create(self, pool_name: str, num_instances: int, instance_family: str):
+        return self._execute_query(
+            f"""\
+            CREATE COMPUTE POOL {pool_name}
+            MIN_NODES = {num_instances}
+            MAX_NODES = {num_instances}
+            INSTANCE_FAMILY = {instance_family};
+        """
+        )
+
+    def show(self):
+        return self._execute_query("show compute pools;")
+
+    def drop(
+        self,
+        pool_name: str,
+    ):
+        return self._execute_query(f"drop compute pool {pool_name};")
+
+    def stop(self, pool_name: str):
+        return self._execute_query(f"alter compute pool {pool_name} stop all services;")
