+++ src/snowcli/cli/snowpark/jobs/commands.py
@@ -1,5 +1,3 @@
-import hashlib
-import os
 import sys
 from pathlib import Path
 
@@ -8,7 +6,7 @@ from typing import TextIO
 
 from snowcli.cli.common.decorators import global_options
 from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
-from snowcli.cli.common.sql_execution import SqlExecutionMixin
+from snowcli.cli.snowpark.jobs.manager import JobManager
 from snowcli.cli.stage.manager import StageManager
 from snowcli.output.decorators import with_output
 
@@ -28,34 +26,6 @@ else:
     BLUE = ""
     GRAY = ""
     ENDC = ""
-
-
-class JobManager(SqlExecutionMixin):
-    def create(self, compute_pool: str, spec_path: Path, stage: str):
-        spec_filename = os.path.basename(spec_path)
-        file_hash = hashlib.md5(open(spec_path, "rb").read()).hexdigest()
-        stage_dir = os.path.join("jobs", file_hash)
-        return self._execute_query(
-            f"""\
-        EXECUTE SERVICE
-        COMPUTE_POOL =  {compute_pool}
-        spec=@{stage}/{stage_dir}/{spec_filename};
-        """
-        )
-
-    def desc(self, job_name: str):
-        return self._execute_query(f"desc service {job_name}")
-
-    def status(self, job_name: str):
-        return self._execute_query(f"CALL SYSTEM$GET_JOB_STATUS('{job_name}')")
-
-    def drop(self, job_name: str):
-        return self._execute_query(f"CALL SYSTEM$CANCEL_JOB('{job_name}')")
-
-    def logs(self, job_name: str, container_name: str):
-        return self._execute_query(
-            f"call SYSTEM$GET_JOB_LOGS('{job_name}', '{container_name}')"
-        )
 
 
 @app.command()
